refactor to rewrite_ssd
