+++ test/gpu/ssd_rewrite.cpp
@@ -27,8 +27,12 @@
 #include <migraphx/program.hpp>
 #include <migraphx/register_target.hpp>
 #include <migraphx/verify.hpp>
+#include <migraphx/gpu/rewrite_ssd.hpp>
+#include <migraphx/dead_code_elimination.hpp>
+#include <migraphx/pass_manager.hpp>
 
 #include <test.hpp>
+#include <limits>
 
 static std::vector<float> run_gpu_topk_values(migraphx::program p,
                                               const migraphx::parameter_map& host_params)
@@ -54,8 +58,8 @@ static std::vector<float> run_gpu_topk_values(migraphx::program p,
     return values;
 }
 
-// Builds the SSD-style nonmaxsuppression -> gather(scores) -> topk chain that
-// simplify_dyn_ops::find_nms_gather_topk rewrites. NMS output is static [max_boxes, 3] with only
+// Builds the SSD-style nonmaxsuppression -> gather(scores) -> topk chain that gpu::rewrite_ssd
+// rewrites. NMS output is static [max_boxes, 3] with only
 // num_selected valid rows. The padded rows point at box 0, so the gather pulls box 0's (highest)
 // score into every padded slot. Without the where-mask rewrite the topk returns duplicates of
 // box 0's score instead of the real detections. Data/gold were generated by using onnxruntime.
@@ -125,6 +129,214 @@ TEST_CASE(test_ssd_rewrite)
     // padded row.
     std::vector<float> gold = {0.9f, 0.8f, 0.7f};
     EXPECT(migraphx::verify::verify_rms_range(values, gold));
+}
+
+static void run_pass(migraphx::module& m)
+{
+    migraphx::run_passes(m, {migraphx::gpu::rewrite_ssd{}, migraphx::dead_code_elimination{}});
+}
+
+// Build boxes/scores -> nonmaxsuppression -> dynamic slice -> gather scores -> topk in `m`
+// (matching what the ONNX NMS parser emits), returning the gather so callers can extend the graph.
+static migraphx::instruction_ref make_nms_gather_topk(migraphx::module& m, bool largest)
+{
+    migraphx::shape boxes_s{migraphx::shape::float_type, {1, 4, 4}};
+    migraphx::shape scores_s{migraphx::shape::float_type, {1, 1, 4}};
+    auto boxes     = m.add_parameter("boxes", boxes_s);
+    auto scores    = m.add_parameter("scores", scores_s);
+    auto max_out   = m.add_literal(int64_t{4});
+    auto iou       = m.add_literal(0.5f);
+    auto score_thr = m.add_literal(0.0f);
+    auto nms       = m.add_instruction(
+        migraphx::make_op("nonmaxsuppression"), boxes, scores, max_out, iou, score_thr);
+    auto indices      = m.add_instruction(migraphx::make_op("get_tuple_elem", {{"index", 0}}), nms);
+    auto num_selected = m.add_instruction(migraphx::make_op("get_tuple_elem", {{"index", 1}}), nms);
+    auto selected     = m.add_instruction(
+        migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}}), indices, num_selected);
+    auto box_col = m.add_instruction(
+        migraphx::make_op("slice", {{"axes", {1}}, {"starts", {2}}, {"ends", {3}}}), selected);
+    auto box_col_1d  = m.add_instruction(migraphx::make_op("squeeze", {{"axes", {1}}}), box_col);
+    auto scores_flat = m.add_instruction(migraphx::make_op("squeeze", {{"axes", {0, 1}}}), scores);
+    auto gather =
+        m.add_instruction(migraphx::make_op("gather", {{"axis", 0}}), scores_flat, box_col_1d);
+    auto topk = m.add_instruction(
+        migraphx::make_op("topk", {{"k", 4}, {"axis", 0}, {"largest", largest ? 1 : 0}}), gather);
+    auto tv = m.add_instruction(migraphx::make_op("get_tuple_elem", {{"index", 0}}), topk);
+    m.add_return({tv});
+    return gather;
+}
+
+// The graph gpu::rewrite_ssd is expected to produce for make_nms_gather_topk: the dynamic slice is
+// gone and the gathered scores are masked so the padded rows become a sentinel. Insertion order
+// mirrors the pass so the printed IR (compared by module equality) matches.
+static void make_nms_gather_topk_masked(migraphx::module& m, bool largest)
+{
+    migraphx::shape boxes_s{migraphx::shape::float_type, {1, 4, 4}};
+    migraphx::shape scores_s{migraphx::shape::float_type, {1, 1, 4}};
+    auto boxes     = m.add_parameter("boxes", boxes_s);
+    auto scores    = m.add_parameter("scores", scores_s);
+    auto max_out   = m.add_literal(int64_t{4});
+    auto iou       = m.add_literal(0.5f);
+    auto score_thr = m.add_literal(0.0f);
+    auto iota_lit  = m.add_literal(
+        migraphx::literal{migraphx::shape{migraphx::shape::int64_type, {4}}, {0, 1, 2, 3}});
+    auto sentinel_val =
+        largest ? std::numeric_limits<float>::lowest() : std::numeric_limits<float>::max();
+    auto sentinel = m.add_literal(
+        migraphx::literal{migraphx::shape{migraphx::shape::float_type, {1}}, {sentinel_val}});
+    auto nms = m.add_instruction(
+        migraphx::make_op("nonmaxsuppression"), boxes, scores, max_out, iou, score_thr);
+    auto indices      = m.add_instruction(migraphx::make_op("get_tuple_elem", {{"index", 0}}), nms);
+    auto num_selected = m.add_instruction(migraphx::make_op("get_tuple_elem", {{"index", 1}}), nms);
+    auto box_col      = m.add_instruction(
+        migraphx::make_op("slice", {{"axes", {1}}, {"starts", {2}}, {"ends", {3}}}), indices);
+    auto box_col_1d  = m.add_instruction(migraphx::make_op("squeeze", {{"axes", {1}}}), box_col);
+    auto scores_flat = m.add_instruction(migraphx::make_op("squeeze", {{"axes", {0, 1}}}), scores);
+    auto gather =
+        m.add_instruction(migraphx::make_op("gather", {{"axis", 0}}), scores_flat, box_col_1d);
+    auto num_selected_bc =
+        m.add_instruction(migraphx::make_op("multibroadcast", {{"out_lens", {4}}}), num_selected);
+    auto less = m.add_instruction(migraphx::make_op("less"), iota_lit, num_selected_bc);
+    auto mask = m.add_instruction(
+        migraphx::make_op("convert", {{"target_type", migraphx::shape::bool_type}}), less);
+    auto sentinel_bc =
+        m.add_instruction(migraphx::make_op("multibroadcast", {{"out_lens", {4}}}), sentinel);
+    auto masked = m.add_instruction(migraphx::make_op("where"), mask, gather, sentinel_bc);
+    auto topk   = m.add_instruction(
+        migraphx::make_op("topk", {{"k", 4}, {"axis", 0}, {"largest", largest ? 1 : 0}}), masked);
+    auto tv = m.add_instruction(migraphx::make_op("get_tuple_elem", {{"index", 0}}), topk);
+    m.add_return({tv});
+}
+
+TEST_CASE(nms_gather_topk_largest)
+{
+    migraphx::module m1;
+    make_nms_gather_topk(m1, true);
+    run_pass(m1);
+
+    migraphx::module m2;
+    make_nms_gather_topk_masked(m2, true);
+
+    EXPECT(m1 == m2);
+}
+
+TEST_CASE(nms_gather_topk_smallest)
+{
+    migraphx::module m1;
+    make_nms_gather_topk(m1, false);
+    run_pass(m1);
+
+    migraphx::module m2;
+    make_nms_gather_topk_masked(m2, false);
+
+    EXPECT(m1 == m2);
+}
+
+// A second dynamic NMS slice in the ancestor cone makes it ambiguous, so the rewrite must not fire.
+TEST_CASE(nms_gather_topk_two_nms_no_change)
+{
+    migraphx::module m1;
+    auto gather = make_nms_gather_topk(m1, true);
+    migraphx::shape boxes_s{migraphx::shape::float_type, {1, 4, 4}};
+    migraphx::shape scores_s{migraphx::shape::float_type, {1, 1, 4}};
+    auto boxes2  = m1.add_parameter("boxes2", boxes_s);
+    auto scores2 = m1.add_parameter("scores2", scores_s);
+    auto nms2    = m1.add_instruction(
+        migraphx::make_op("nonmaxsuppression"), boxes2, scores2, m1.add_literal(int64_t{4}));
+    auto indices2 = m1.add_instruction(migraphx::make_op("get_tuple_elem", {{"index", 0}}), nms2);
+    auto num_selected2 =
+        m1.add_instruction(migraphx::make_op("get_tuple_elem", {{"index", 1}}), nms2);
+    auto selected2 = m1.add_instruction(
+        migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}}), indices2, num_selected2);
+    auto box_col2 = m1.add_instruction(
+        migraphx::make_op("slice", {{"axes", {1}}, {"starts", {2}}, {"ends", {3}}}), selected2);
+    auto box_col2_1d = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {1}}}), box_col2);
+    // Mask R-CNN concatenates the selected box indices from multiple NMS ops before the final
+    // gather/topk; two NMS slices in the ancestor cone must leave the pass a no-op.
+    auto gather_idx   = gather->inputs().at(1);
+    auto concatenated = m1.insert_instruction(
+        gather, migraphx::make_op("concat", {{"axis", 0}}), gather_idx, box_col2_1d);
+    m1.replace_instruction(gather, gather->get_operator(), gather->inputs().at(0), concatenated);
+
+    migraphx::module m2 = m1;
+    run_pass(m1);
+
+    EXPECT(m1 == m2);
+}
+
+// Full SSD output tail: parse_topk's var-k slices feed downstream gathers. rewrite_ssd should set
+// the topk k to the cap (min-with-dim literal = 3), keep the gathers static, and trim only the
+// output to min(cap, num_selected). Distinct dims (max_boxes=5, coords=4, cap=3) keep the output
+// detection axis unambiguous.
+TEST_CASE(nms_gather_topk_static_outputs)
+{
+    migraphx::module m1;
+    migraphx::shape boxes_s{migraphx::shape::float_type, {1, 5, 4}};
+    migraphx::shape scores_s{migraphx::shape::float_type, {1, 1, 5}};
+    auto boxes     = m1.add_parameter("boxes", boxes_s);
+    auto scores    = m1.add_parameter("scores", scores_s);
+    auto max_out   = m1.add_literal(int64_t{5});
+    auto iou       = m1.add_literal(0.5f);
+    auto score_thr = m1.add_literal(0.0f);
+    auto nms       = m1.add_instruction(
+        migraphx::make_op("nonmaxsuppression"), boxes, scores, max_out, iou, score_thr);
+    auto indices = m1.add_instruction(migraphx::make_op("get_tuple_elem", {{"index", 0}}), nms);
+    auto num_selected =
+        m1.add_instruction(migraphx::make_op("get_tuple_elem", {{"index", 1}}), nms);
+    auto selected = m1.add_instruction(
+        migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}}), indices, num_selected);
+    auto box_col = m1.add_instruction(
+        migraphx::make_op("slice", {{"axes", {1}}, {"starts", {2}}, {"ends", {3}}}), selected);
+    auto box_col_1d  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {1}}}), box_col);
+    auto scores_flat = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0, 1}}}), scores);
+    auto gathered =
+        m1.add_instruction(migraphx::make_op("gather", {{"axis", 0}}), scores_flat, box_col_1d);
+    auto topk = m1.add_instruction(
+        migraphx::make_op("topk", {{"k", 5}, {"axis", 0}, {"largest", 1}}), gathered);
+    auto topk_ind = m1.add_instruction(migraphx::make_op("get_tuple_elem", {{"index", 1}}), topk);
+    // parse_topk var-k: k = min(cap=3, dim = elements of gathered scores)
+    auto dim = m1.add_instruction(migraphx::make_op("dimensions_of", {{"start", 0}, {"end", 1}}),
+                                  gathered);
+    auto cap_lit =
+        m1.add_literal(migraphx::literal{migraphx::shape{migraphx::shape::int64_type, {1}}, {3}});
+    auto kcat       = m1.add_instruction(migraphx::make_op("concat", {{"axis", 0}}), cap_lit, dim);
+    auto k          = m1.add_instruction(migraphx::make_op("reduce_min", {{"axes", {0}}}), kcat);
+    auto sliced_ind = m1.add_instruction(
+        migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}}), topk_ind, k);
+    // downstream: gather the box coordinates and add a batch dim
+    auto boxes_table = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), boxes);
+    auto det_boxes =
+        m1.add_instruction(migraphx::make_op("gather", {{"axis", 0}}), boxes_table, sliced_ind);
+    auto out = m1.add_instruction(migraphx::make_op("unsqueeze", {{"axes", {0}}}), det_boxes);
+    m1.add_return({out});
+
+    run_pass(m1);
+
+    // Only the output is dynamic: it is a slice trimming to min(cap, num_selected).
+    auto ret = std::prev(m1.end());
+    EXPECT(ret->name() == "@return");
+    auto out_slice = ret->inputs().at(0);
+    EXPECT(out_slice->name() == "slice");
+    EXPECT(out_slice->get_shape().dynamic());
+
+    auto new_k = out_slice->inputs().at(1);
+    EXPECT(new_k->name() == "min");
+    EXPECT(new_k->inputs().at(0)->name() == "@literal"); // cap
+    auto num_sel = new_k->inputs().at(1);                // num_selected from the NMS
+    EXPECT(num_sel->name() == "get_tuple_elem");
+    EXPECT(num_sel->inputs().at(0)->name() == "nonmaxsuppression");
+
+    // Everything feeding the trim is static (static until the outputs).
+    auto unsq = out_slice->inputs().at(0);
+    EXPECT(not unsq->get_shape().dynamic());
+
+    // The TopK now sorts only cap = 3 (the maximum k), on a static input.
+    auto det_gather = unsq->inputs().at(0);
+    auto topk_gte   = det_gather->inputs().at(1);
+    auto topk_new   = topk_gte->inputs().at(0);
+    EXPECT(topk_new->name() == "topk");
+    EXPECT(topk_new->get_operator().to_value().at("k").to<int>() == 3);
+    EXPECT(not topk_new->inputs().at(0)->get_shape().dynamic());
 }
 
 int main(int argc, const char* argv[]) { test::run(argc, argv); }
